python-client/requirements.txt
aiohttp
python-socketio
adbutils